Move Annotation form types to Core
