+++ src/Wallabag/CoreBundle/Form/Type/EditAnnotationType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wallabag\AnnotationBundle\Form;
+namespace Wallabag\CoreBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
